+++ s_sound.h
@@ -46,7 +46,7 @@
 // Sets channels, SFX and music volume,
 //  allocates channel buffer, sets S_sfx lookup.
 //
-void S_Init(int32_t sfxVolume, int32_t musicVolume);
+void S_Init(int16_t sfxVolume, int16_t musicVolume);
 
 // Kills all sounds
 void S_Stop(void);
@@ -82,8 +82,8 @@ void S_ChangeMusic(musicenum_t music_id, boolean looping);
 // Updates music & sounds
 //
 void S_UpdateSounds(void);
-void S_SetMusicVolume(int32_t volume);
-void S_SetSfxVolume(int32_t volume);
+void S_SetMusicVolume(int16_t volume);
+void S_SetSfxVolume(int16_t volume);
 
 
 #endif
